+++ src/__tests__/element.ts
@@ -153,3 +153,26 @@ describe('#equal()', () => {
     expect(nested_element_1.equal(nested_element_4, true)).toBe(false);
   });
 });
+
+describe('#has()', () => {
+  const required = 'required';
+  const normal_element = new NormalElement({required});
+  const nested_element = new NestedElement({
+    element: normal_element,
+  });
+  const nested_nested_element = new NestedElement({
+    element: new NestedElement({
+      element: normal_element,
+    }),
+  });
+  it('should just check its own parameters while is_deep_has = false', () => {
+    expect(normal_element.has(normal_element, false)).toBe(false);
+    expect(nested_element.has(normal_element, false)).toBe(true);
+    expect(nested_nested_element.has(normal_element, false)).toBe(false);
+  });
+  it('should check all its parameters and sub-parameters while is_deep_has = true', () => {
+    expect(normal_element.has(normal_element, true)).toBe(false);
+    expect(nested_element.has(normal_element, true)).toBe(true);
+    expect(nested_nested_element.has(normal_element, true)).toBe(true);
+  });
+});